make export of HeaderTabs consistent with other components export
